delete sales man alert dialog

+++ lib/views/screens/home_screen/admin_home_screen/admin_home_screen_widgets/admin_home_screen_body_widget/drivers_details_widget/driver_details_holder/salesman_details_holder/available_details_widget/address_holder_widget/delete_address_widget/delete_address_widget.dart
@@ -1,4 +1,5 @@
 import 'package:farm_dairy/controllers/admin_home_screen_controllers.dart';
+import 'package:farm_dairy/models/common_widgets/text_widget.dart';
 import 'package:flutter/material.dart';
 
 class DeleteAddressWidget extends StatelessWidget {
@@ -6,7 +7,7 @@ class DeleteAddressWidget extends StatelessWidget {
     super.key,
     required this.screenSize,
     required this.isDarkMode,
-    required this.documentId
+    required this.documentId,
   });
 
   final Size screenSize;
@@ -17,13 +18,60 @@ class DeleteAddressWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
       onPressed: () {
-        deleteSalesMan(documentId: documentId, context: context);
-      }, 
+        showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: TextWidget(
+                text: 'Confirm Deletion',
+                color: isDarkMode ? Colors.white : Colors.black,
+                size: screenSize.width / 20,
+                fontFamily: 'Roboto', // Replace with your font family if required
+                weight: FontWeight.bold,
+              ),
+              content: TextWidget(
+                text: 'Are you sure you want to delete this address?',
+                color: isDarkMode ? Colors.white70 : Colors.black87,
+                size: screenSize.width / 25,
+                fontFamily: 'Roboto',
+                weight: FontWeight.normal,
+              ),
+              actions: [
+                TextButton(
+                  onPressed: () {
+                    Navigator.of(context).pop(); // Close the dialog
+                  },
+                  child: TextWidget(
+                    text: 'Cancel',
+                    color: isDarkMode ? Colors.white : Colors.black,
+                    size: screenSize.width / 30,
+                    fontFamily: 'Roboto',
+                    weight: FontWeight.normal,
+                  ),
+                ),
+                TextButton(
+                  onPressed: () {
+                    deleteSalesMan(documentId: documentId, context: context);
+                    Navigator.of(context).pop(); // Close the dialog after deletion
+                  },
+                  child: TextWidget(
+                    text: 'Delete',
+                    color: Colors.red,
+                    size: screenSize.width / 30,
+                    fontFamily: 'Roboto',
+                    weight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            );
+          },
+        );
+      },
       icon: Icon(
         Icons.delete,
-        size: screenSize.width/25,
-        color: isDarkMode?Colors.white:Colors.black,
-      )
+        size: screenSize.width / 25,
+        color: isDarkMode ? Colors.white : Colors.black,
+      ),
     );
   }
 }
